Consolidate etl user functions
